==> analysis/03_Duncan_rings-to-pith.Rmd <==
---
title: "Modeling Uncertainty in Pith Offsets"
author: 
  - Kyla S. Zaret, Portland State University
  - Eric R. Buhle, Biomark Applied Biological Services and Northwest Fisheries Science Center
  - Andrés Holz, Portland State University
date: "`r format(Sys.time(), '%Y-%m-%d')`"
output: 
  html_document:
    df_print: paged
    fig_caption: true
    theme: cerulean
    toc: true
    toc_float: true
---

<style type="text/css">
  body{
  font-size: 11pt;
}
</style>

```{r Rmd_setup, include=FALSE}
knitr::opts_chunk$set(tidy = FALSE, highlight = TRUE, comment = NA, 
                      dev = "png", dev.args = list(type = "cairo-png"), dpi = 300,
                      out.width = "60%", fig.align = "center")

library(here)
if(!require(captioner))
  devtools::install_github("adletaw/captioner")
library(captioner)
fig_nums <- captioner("Figure ", suffix = ": ", auto_space = FALSE, style = "b", style_prefix = TRUE)
```
```{r width, include=FALSE}
options(width = 130)
```
```{r read_chunks, echo = FALSE}
knitr::read_chunk(here("analysis","03_Duncan_rings-to-pith.R"))
```

## Setup

```{r setup, message=FALSE}
```
```{r load_stanmodels, echo=FALSE}
if(file.exists(here("analysis","results","duncan_rings-to-pith.RData")))
   load(here("analysis","results","duncan_rings-to-pith.RData"))
```

## Overview

The second component of ageing error arises from cores that missed the chronological center of the stem. In such cases it is necessary to estimate the pith offset, or the number of rings between the innermost observable ring and the pith. Duncan (1989) developed a geometric formula to calculate the pith offset, assuming concentric circular rings. This method first calculates the missing radius $r$ given the measured length $L$ and height $h$ of the largest fully visible arc formed by the innermost growth rings:

$$
r = \dfrac{L^2}{8h} + \dfrac{h}{2}.
$$

The mean width of the innermost $N_{d}$ visible rings, $\bar{d} = (1 / N_d) \sum_{i=1}^{N_{d}}{d_i}$, is then used as an estimate of the mean unobserved ring width to calculate missing rings as $r / \bar{d}$.

Propagating uncertainty blahblah...

## Data Collection

Describe sampling protocol, criteria for excluding cores, etc...

```{r data}
```

Inspect data objects...

## Hierarchical Linear Model

Describe modeling approach...

$$
\begin{aligned}
\text{log}(d_i) &\sim N(\mu_{\text{tree}[i]}, \sigma_d) \\
\mu_{\text{tree}[i]} &= \beta_{\text{patch}[i]} + b_{\text{tree}[i]} \\
b_{\text{tree}} &\sim N(0,\sigma_\text{tree}) \\
\end{aligned}
$$


```{r fit_lmer, eval = !exists("duncan_lmer")}
```
```{r print_fit_lmer}
```

### Posterior Predictive Checking

Discuss posterior predictive checking...

```{r sim_ppd}
```

Posterior predictive marginal density...

```{r ppc_dens_overlay_patch, echo=FALSE, fig.width=7, fig.height=7}
<<ppc_dens_overlay_patch>>
```

Posterior predictive check for mean log ring width...

```{r ppc_mean_patch, echo=FALSE, message=FALSE, fig.width=7, fig.height=7}
<<ppc_mean_patch>>
```

Posterior predictive check for SD of log ring width...

```{r ppc_sd_patch, echo=FALSE, message=FALSE, fig.width=7, fig.height=7}
<<ppc_sd_patch>>
```

Normal QQ plot of tree-level intercepts, grouped by patch [currently broken because of some mismatch between trees in `inner_rings` and trees in `duncan`]...

```{r qqnorm_intercept_patch, eval=FALSE, fig.height=7, fig.width=7, include=FALSE}
<<qqnorm_intercept_patch>>
```

Normal QQ plot of observation-level residuals, grouped by patch...

```{r qqnorm_resid_grouped, echo=FALSE, fig.width=7, fig.height=7}
<<qqnorm_resid_patch>>
```


## Pith Offsets

We transformed the posterior samples to calculate the posterior distribution of median inner ring width for each tree, $\widehat{d_{\text{tree}}} = \text{exp}(\mu_{\text{tree}})$, which we then used in place of the sample mean $\bar{d}$ to estimate the missing rings to pith as $r / \widehat{d_{\text{tree}}}$. We used the median rather than the mean as a more robust measure of location for the lognormal data distribution, although in this case the difference is slight.

```{r calc_rtp}
```


```{r duncan_rings-to-pith_joyplots, echo=FALSE, message=FALSE, fig.width=7, fig.height=7}
<<duncan_rings-to-pith_joyplots>>
```
